+++ openradar/config.py
@@ -132,6 +132,9 @@ WAIT_EXPIRE_DELTA = datetime.timedelta(minutes=3)
 # Productcopy settings, for fews import, for example.
 COPY_TARGET_DIRS = []
 
+# Root path to opendap data, for retrieval only.
+OPENDAP_ROOT = 'http://opendap.nationaleregenradar.nl/thredds/dodsC/radar'
+
 # FTP settings
 # Publishing
 FTP_HOST = ''  # Empty to skip ftp publishing